+++ controllers/beneficiary.controller.js
@@ -119,7 +119,7 @@ exports.updateBeneficiary = async (req, response) => {
             fechaIngreso,
             foto,
             estatus,
-            idDiscapacidad
+            discapacidades
         } = req.body;
 
         const result = await Beneficiary.update(
@@ -135,7 +135,7 @@ exports.updateBeneficiary = async (req, response) => {
             fechaIngreso !== undefined ? fechaIngreso : existingData.fechaIngreso,
             foto !== undefined ? foto : existingData.foto,
             estatus !== undefined ? estatus : existingData.estatus,
-            idDiscapacidad !== undefined ? idDiscapacidad : existingData.idDiscapacidad,
+            discapacidades !== undefined ? discapacidades : existingData.discapacidades,
         );
         
         response.status(200).json({
@@ -154,9 +154,8 @@ exports.updateBeneficiary = async (req, response) => {
             ...(fechaIngreso && { fechaIngreso }),
             ...(foto && { foto }),
             ...(estatus !== undefined && { estatus }),
-            ...(idDiscapacidad !== undefined && { idDiscapacidad })
+            ...(discapacidades !== undefined && { discapacidades })
             },
-            affectedRows: result[0]?.affectedRows || result.affectedRows
         }
         });