+++ models/disabilities.model.js
@@ -1,7 +1,6 @@
 const database = require('../utils/db.js');
 module.exports = class Disability {
-    constructor(idDisability, name, description) {
-        this.idDisability = idDisability;
+    constructor(name, description) {
         this.name = name;
         this.description = description;
     }
@@ -13,4 +12,38 @@ module.exports = class Disability {
         const rows = await database.query(sql);
         return rows;
     }
+
+    async updateById(idDiscapacidad) {
+        return database
+          .execute(
+            `UPDATE ListaDiscapacidades
+            SET
+              nombre = ?,
+              caracteristicas = ?
+            WHERE idDiscapacidad = ?`,
+            [
+              this.name,
+              this.description,
+              idDiscapacidad
+            ]
+          )
+          .then(() => {
+            return database.execute(
+              `SELECT *
+              FROM ListaDiscapacidades
+              WHERE idDiscapacidad = ?`, [idDiscapacidad]
+            );
+          });
+      }
+
+      static async fetchDisability(id) {
+        const query = `
+            SELECT *
+            FROM ListaDiscapacidades
+            WHERE idDiscapacidad = ?
+        `;
+
+        const rows = await database.execute(query, [id]);
+        return rows.length > 0 ? rows[0] : null;
+      }
 }