+++ routes/disabilities.routes.js
@@ -6,4 +6,9 @@ const disabilityController = require('../controllers/disabilities.controller');
 // Ruta para BEN-001
 router.get('/list', disabilityController.fetchAll);
 
+// Ruta para modificar discapacidad (US CAP-003)
+router.post("/update", disabilityController.updateDisabilities);
+
+router.get("/:idDiscapacidad", disabilityController.fetchDisability);
+
 module.exports = router;