+++ routes/upload.routes.js
@@ -1,0 +1,33 @@
+const express = require('express');
+const multer  = require('multer');
+const path = require('path');
+const fs = require('fs'); 
+const router = express.Router();
+const storage = multer.diskStorage({
+  destination: function(req, file, cb){
+    cb(null, 'uploads/'); 
+  },
+  filename: function (req, file, cb) {
+    const unique = Date.now() + "-" + Math.round(Math.random() * 1E9);
+    cb(null, unique + path.extname(file.originalname));
+  }
+});
+
+const upload = multer({ storage });
+router.post('/', upload.single("image"), (req, res) => {
+  if (!req.file) {
+    return res.status(400).json({ success: false, message: "No se envió archivo." });
+  }
+
+  const fileUrl = `${req.protocol}://${req.get('host')}/uploads/${req.file.filename}`;
+
+  res.json({
+    success: true,
+    url: fileUrl,
+    filename: req.file.filename
+  });
+});
+
+
+
+module.exports = router;